bomx: allow hdr metadata use in non-tunneled mode.
initially we restricted hdr metadate to be only applicable to tunnel
decoder simply becuase that is what was defined from the specification.

support for metadata in non-tunneled decoder is unspecified and as such
left to the implementation to determine usage.  since we can use that
information regardless of the tunnel mode selection, we can lift the
initial restriction.

this does not mean that one should not strive to use tunnel mode for
hdr streams in general since performance on those would be much better,
but it gives some amount of flexibility.

refs #SWANDROID-4340

Change-Id: I69c3990c79e236fb07875f934610970d86d13b2b
Signed-off-by: Pierre Couillaud <pierre.couillaud@broadcom.com>