Improving buffer lifecycle model for video decoder
This change allows the omx layer to maintain ownership of frames after they
are sent for display and also allows frames to be returned out of order to
nexus.  This should allow for destriping on-demand and could re-enable
metadata output mode if desired.  The only case that still requires buffers
to be invalidated in omx is a flush or decoder stop.  In those cases, the
decoder internally reclaims all the memory and this can't be avoided.

refs #SWANDROID-1760

Change-Id: I5b0fc65613675acbf0775e30402ad22d360e31c6