hwc2: clear out cached layer fence upon pushed to composition.
in some circumstances, SF will push composition to display without
resetting the layer buffer and fence content; when this happens and
if the prior layer fence had been closed already, we may end up managing
a fence that has become stale.  to avoid this condition we reset the
fence as soon as we push a composition unit to the composer loop, this
ensures that the fence will be acted on properly and if the fence is
not updated in the next composition, it would appear as if the fence
for the layer has timed out already and the buffer can be accessed
immediately, which happens to be handled properly because if the buffer
has been removed we would detect and abort the composition and if the
buffer is still present, we would just re-issue the same composition.

refs #SWANDROID-4389

Change-Id: I9c72ad80b998536767f262d9d5b0c2800f89635e
Signed-off-by: Pierre Couillaud <pierre@broadcom.com>