nxserver: expand scope of memory tracker to include nxmem and bcmpip.
nxmem are allocations which come from the nexus ashmem driver, this includes
framebuffers and allocations from generic usage outside of graphics stack; or
allocations from graphics stack directed to nxmem where applicable.

bcmpip are allocations specific to graphics stack per process, though do not
need to be restricted to those, this is the only user of the bcmpip interface
for the time being.

properly make decisions on memory allocations using all the sources possible:
mmu (graphics only), nxmem and bcmpip (as documented above).

this allows to apply the memory tracking of rmlmk to all device configuration
including those without mmu or for which mmu is disabled for other reasons.

refs #SWANDROID-4610

Change-Id: I1bf5ed5cd8c4ee2d2ffae907e45a281a5a8b7e44
Signed-off-by: Pierre Couillaud <pierre.couillaud@broadcom.com>