wol: move control over wake-on-lan setup in inexus for treble.
the power hal is not capable of controlling socket directly due to
policy restrictions, so we move the processing to inexus which can
since the wol setup is fairely isolated.  the power hal is still in
charge and control over it, but the action is deferred to the vendor
hidl service.

refs #SWANDROID-4438

Change-Id: Ifa76c960970a220f51d2f7f7d5ce55db95af4e4d
Signed-off-by: Pierre Couillaud <pierre@broadcom.com>
(cherry picked from commit a76819ce454d6f51ba413ecd55c4587ad33fb373)