gralloc: allow allocations of pure sw video buffers.
those buffers are used purely in a sw fashion and are not directly
expected to be used for hw texturing either, typically used for pass
through camera type applications.

we allow them to be created but we still limit their sizes to the
maximum hardware plane optimal sizes.  the main usage for those in the
context of ATV are mostly to pass certification on some odd tests in the
camera group (should not even run for ATV).

refs #SWANDROID-4685

Change-Id: I52a08bb30a0d5713b80431fbe5bfc0d08eae99ac
Signed-off-by: Pierre Couillaud <pierre@broadcom.com>
(cherry picked from commit c3f1caab3915a329baad693968e7f5cd74554b33)