vendor: add support for virtualized decoder (aka mosaic mode).
adding ability to use the main decoder as virtualized, which allows creation
of x (lower) decoders of capability maximale y.  we do not allow virtualization
of the pip decoder at the moment (which when available is still exposed as the
"redux" decoder).

x and y depend on the device and the rts mode in use and is evaluated run time.
in the common usage model, x is 3 and y is 1080p.

once main decoder is virtualized, new instances created must be virtualized as well
(this is validated and enforced automatically); in order to recuperate the initial
capabilities of the main decoder (eg 4k) all virtualized decoder must be terminated.
it is the responsibility of the user application to manage this life cycle.

in this initial version, to signal need to virutalize we use a dynamic property.
ongoing investigation in finding a different|better way that works with the omx|framework
flow.

refs #SWANDROID-4636

Change-Id: I29bda853a572926d01a4d714eb54d9e832ea6525
Signed-off-by: Pierre Couillaud <pierre@broadcom.com>