hwc2: preserve display configuration across android query lifetime.
this is required to ensure the information at the framework is always the
one queried by it; regardless of the device connectivity changes underneath (as
example hotplug).  this can be seen as a limitation; in fact resulting from
that android framework does not like display-0 to be hotplug (we never report
it).  when that gets addressed, we can relax this limitation.

in practice, locking the configuration at android query is the proper thing
to do, but this may fail some odd cases where a different endpoint with complete
different feature set is hotplug and then framework tries to swap configuration
on it for some unsupported one (this would fail)...  all in all however this is
quite far fetched production situation, so we should be fine for now.

refs #SWANDROID-XXXX

Change-Id: I69af50d48feabfb02d99466390185c5e98353d29
Signed-off-by: Pierre Couillaud <pierre.couillaud@broadcom.com>
(cherry picked from commit 5dc4345602a3dee4913d8e7a731a0757461b8944)