hwc2: preserve display configuration across android query lifetime.
note: backport from O to N.

this is required to ensure the information at the framework is always the
one queried by it; regardless of the device connectivity changes underneath (as
example hotplug).  this can be seen as a limitation; in fact resulting from
that android framework does not like display-0 to be hotplug (we never report
it).  when that gets addressed, we can relax this limitation.

in practice, locking the configuration at android query is the proper thing
to do, but this may fail some odd cases where a different endpoint with complete
different feature set is hotplug and then framework tries to swap configuration
on it for some unsupported one (this would fail)...  all in all however this is
quite far fetched production situation, so we should be fine for now.

refs #SWANDROID-4294

Change-Id: I37ca93737820aa6303a5ea9b26ffbf79b708b418
Signed-off-by: Pierre Couillaud <pierre.couillaud@broadcom.com>