boot_control: delay marking of successful boot until we can validate middleware.
this allows to avoid marking an image|slot valid if the nexus middleware is
actually not proper (for whatever reason).  we now check that nexus has
booted and can be talked to prior to makring success, which allows to filter
out cases where nexus is not working; this may not avoid all bad classifications
but it should give a fairly good baseline.

refs #SWANDROID-4772

Change-Id: I0868fc1c74b8b4b52f157c8976e81464fe1e9574
Signed-off-by: Pierre Couillaud <pierre.couillaud@broadcom.com>
(cherry picked from commit 8f4b25c77131412880987aa122518777b0dbf495)