boot_control: delay marking of successful boot until we can validate middleware.
this allows to avoid marking an image|slot valid if the nexus middleware is
actually not proper (for whatever reason).  we now check that nexus has
booted and can be talked to prior to makring success, which allows to filter
out cases where nexus is not working; this may not avoid all bad classifications
but it should give a fairly good baseline.

refs #SWANDROID-4772

Change-Id: I151e5cee0e57de0bd63ec0e5bea4728ba78f3293
Signed-off-by: Pierre Couillaud <pierre.couillaud@broadcom.com>
(cherry picked from commit d323021f6c89a6ce3987651ac008e01fbff28606)