hals: keymaster.
keymaster2|sage tl: initial integration.

requirements: ssd (rpmb) must be started first (this is handled
by the startup sequence update); km certs must be installed on the
device (in /dev/hwcfg as add on to drm.bin; either generate a new
drm.bin with those, or add a km.bin).

for bcm reference devices, including zd|zb but limited to zeus 4 only
platforms, a fallback mechanism for the lack of km certs is made
available with a default canned certificate - this is only meant to
assist deployment of the feature on existing systems and should not be
considered a proper solution for a production device.

warning: failure to have a properly setup device to run keymaster
may result in loss of functionality including failure to boot android.

refs #SWANDROID-1158

Change-Id: I65fe7eff3c4b999dd62de052c166fdb067b0c685
Signed-off-by: Pierre Couillaud <pierre@broadcom.com>