bomx: cleanup secure input buffers properly
When calling the function 'FreeInputBuffers' during the
cleanup of secure components, make sure the call is made
from the secure component destructor and not from the
destructor of the base class. This is required because
the function 'FreeInputBuffer' is virtual.

refs #SWANDROID-4854

Change-Id: I6123273c5082586af5af511e6d5452e3bc507248
Signed-off-by: Luis Navarro <luis.navarro@broadcom.com>

Conflicts:
	media/libstagefrighthw/bomx_video_decoder.cpp
	media/libstagefrighthw/bomx_video_decoder.h

(cherry picked from commit 12c6b15718d5ddd9d0aab3dbb35d503fc3fbb58b)