bomx: tune down non-supported parameter|extensions warnings to info.
if an index or extension is not supported, there are two cases.

1) the support is not mandatory and the component is working properly without
it, in which case outputing a warning is not necessary (and can in fact be
confusing).  this is the most common case in the current integration.

2) the support is required and not supporting it will fail the component
anyways (android would error out on the playback), in which case outputing an
info still shows that things were not supported to give some clue into the
possible root of the problem.  this could be the case when integrating new
android version or feature which was not supported|enabled before at bomx
level.

either ways, reducing this particular warning to info is not causing any
harm and avoid confusing the automated testing log scanner.

refs #SWANDROID-4896

Change-Id: If05ec7a5836e123ddbf37715931171e1a4168f2f
Signed-off-by: Pierre Couillaud <pierre.couillaud@broadcom.com>
(cherry picked from commit 8bd313ca83443ccf0f492803cec4b64f2b3ee9ef)