bomx|hwc2: video client deregistration and eotf forced control.
allow bomx to signal termination of the video clients it requests
usage to hwc-binder.  when the last video client has terminated,
we use that information to determine the need to forcefully change
the transfer function if applicable on hdr connected display.

at the moment, because we are not supporting|integrating hdr for
graphics, the decision is simple. this allows to control the dynamic
range specifically at end of active playback.

refs #SWANDROID-4904

Change-Id: I9c72ffee6943833a6d33eca45e13c21da11700ec
Signed-off-by: Pierre Couillaud <pierre.couillaud@broadcom.com>
(cherry picked from commit 3c79c513089a81b1d984aca5769126c4c1550127)