nxserver: forcefully disable hdcp if we detect mismatched configuration.
because on some systems (such as low end 7260), the lack of hdcp proper
configuration can lead to cpu exhaustion due to the inherent retry algorithm
of the hdcp identification, we purposefully force hdcp disabling if and when
we think hdcp should be supported (by default reference it should), but we
cannot locate keys for it.  this particular situation happens when validating
the vts environment through the gsi image loading because the root for the
keybox including the hdcp drm's is no longer created for the device (since
it is creates through the /hwcfg root node which is device specific and will
not be part of the gsi root environment).

refs #SWANDROID-4947

Change-Id: Ieda8ddc308765975bb93ca16a7b4f39ee38ac9a7
Signed-off-by: Pierre Couillaud <pierre@broadcom.com>
(cherry picked from commit 3eae6db41a547adacca1832c6f8d0edffb5c50ef)