audio: partial revert of SWANDROID-5176 (do not mute on null volume)
the 'mute on null volume' change causes audio to be permanently muted
when exercising the msd hal integration (such would not fall under the
dolbyMs use case).

as this change is only made in an attempt to satisfy netflix certification
(which appears incomplete anyway), reverting it to restore audio in msd hal
usage model is the proper approach, further work will be needed down the
line to ensure both the netflix requirements and the (new) msd hal
integration can live happily.

silenced a too noisy log (enable very verbose if wanted).

refs #SWANDROID-5176

Change-Id: I46d2e5aca868ab35f94b09fc1c95845e4f356189
Signed-off-by: Pierre Couillaud <pierre@broadcom.com>