sec: oemcrypto: nexus hook up is conditional to not changing android.
for the time being, we cannot use a 'vanilla' widevine android codebase,
we still need to overwrite some functionality in order to properly
integrate what we need.  one of such being the nexus hook up to allow
standby monitoring; such needs to happen at the service level to be able
to act properly on it.

when time comes to use a vanilla android oemcrypto, or for delivery of
prebuilts to google, this flag needs to be flipped however.

refs #SWANDROID-5136

Change-Id: Icee7497efd3b2c137577d8c2a23102b85b7d1861
Signed-off-by: Pierre Couillaud <pierre.couillaud@broadcom.com>