bomx: use block handle to pass secure buffer across processes.
the secure buffer created in the bomx context and used in the drm
context needs to be passed around via memory handle rather than
addresses, since the address is a virtual one and can differ from
process to process.

until now we have not had problems with this because the virtual
address happened to be logically always map to the same range in
all processes by nexus (ie 'fake' address range for crr heap was
fixed to always the same location in all processes), but this is
no longer true in kernel64+android32 mixed mode where the 'fake'
address range is dynamically allocated through /dev/zero mapping
to avoid colliding with other (non 'fake') mappings of a given
process.

passing in the block handle allows each process to properly use
the 'fake' address range dedicated for its use.

refs #SWANDROID-5341

Change-Id: I9224f2d32b0dde011e89773c23e852532b68be09
Signed-off-by: Pierre Couillaud <pierre.couillaud@broadcom.com>