hals: power: remove use of nexus gpio when subsystem not available.
*** this will need some rework eventually.

the nexus compat.h stubs for lack or presence of gpio subsystem are
not working for android as they trigger redefinition errors on types
changed (-Werror) and they do not have the full definition of types
required anyway.

so, instead of trying to use those nexus stubs, compile out the
function missing instead using the appropriate #ifdef (sigh!).

refs #SWANDROID-5347

Change-Id: Iee0a739455ff1500ca2a068073bd53b0f21b0858
Signed-off-by: Pierre Couillaud <pierre.couillaud@broadcom.com>