BOMX: Associate decoder frame buffer with metadata buffer instead
Apparently the decoder frame buffer is associated to the gralloc private
handle instead of the metadata buffer which is provided from
UseBuffer(). The gralloc private handle in a metadata buffer might change
during the output buffer lifecycle so when frameworks releases the
output buffer the reverse lookup would fail to find the corresponding
decoder frame buffer. It leads to some of the decoder frame buffers
being held off from recycling back to NEXUS video decoder.

During the VTS run with the GSI image, this issue is exposed because not
enough decoder frame buffers are recycled back so NEXUS video decoder
stops outputing any more frames.

refs #SWANDROID-5467

Change-Id: Ic5ddb1a227758b9f7f416b8a7f90b6d532b427e4