vendor: consolidate bcm properties, unify naming scheme.
this consolidates all (vendor) bcm properties into a single subtree
to allow better exposure, naming and control.

removed duplication (or n-plication, n>2) of properties across
different realms.

all properties defined this way are classified as "nexus_prop" for
selinux realm. this is distinct from the android's "vendor_default_prop"
realm which implies different limitations.

all properties referenced in the code should use a #define, built as
follow:

     BCM_[RO|DYN|PERSIST][_ACT_|<empty>]_<MODULE-OWNER>_<NAME>

 a) "BCM_" is the common prefix, those properties are all bcm defined.
 b) [RO|DYN|PERSIST]: property type. RO is read only, DYN is dynamic
    properties (can be reset overtime) and PERSIST is persistent
    properties.
 c) [_ACT_|<empty>]: if the property is "actionable", add the _ACT_
    keyword, oterwise leave empty (see below).
 d) MODULE-OWNER is the module this property originally belongs to, it
    should match the *_prop.h header, eg: AUDIO, MEDIA, HWC2.
 e) NAME defines the property name inside the scoped realm.

the actual property string needs to start with a prefix that maps it to
a "nexus_prop", as defined in the sepolicy's property_contexts for the
device.

"nexus_prop" are not compatible with the "actionable" restrictions of
android P, because actionable properties must be "vendor_default_prop"
which implies they cannot be set outside of init|vendor_init.  if you
need an actionable property, you may create or rename as necessary such.

refs #SWANDROID-5504

Change-Id: I2c9e45fb55b2742d1fedc381cca1664efb9ebbf8
Signed-off-by: Pierre Couillaud <pierre.couillaud@broadcom.com>