bomx: cycle through flushed port regardless of the association.
it appears there are cases where the port is updated despite being
used for a video texturing case, this has for effect to reset the
gralloc-to-bomx association without a flush; which should be a problem
for gl compliance (and so v3d expectations).

to help pass certification, we are going to allow that for the time
being.

refs #SWANDROID-5506

Change-Id: Iadb32b88a6a23bf18a7e77a00212d34bc01949ad
Signed-off-by: Pierre Couillaud <pierre.couillaud@broadcom.com>