hals: keymaster: skip km latent ops when in nexus standby.
there may be a window of time during which nexus has placed km
in standby yet android is finishing off sleeps operations of its
own (such as removing temporary certificates).  if such happens,
we need to capture the ops and ignore it properly or we may end
up restoring km state to act on the requested ops which would
consequently block further standby completion underneath nexus
already ack'ed knowledge.

this is a bit racy, but we can only control so much of the process
given android is the one managing the full shutodwn procedure.

refs #SWANDROID-5329

Change-Id: Ide9bb3fa8cc3bf3c5e11c8702e51d4a1b8a6ec21
Signed-off-by: Pierre Couillaud <pierre.couillaud@broadcom.com>