bomx: update of speculative port buffers count on format change.
to speed up the port buffer flow and prevent prolonged video decoder
stall on a port buffer format change causing excessive frame drop
(excessive is in contrast with current drop level during the port
format change), we reduce the number of port buffer which we had
speculatively raised at some other point in time for allowing proper
video texturing scenario to work (another vdec constraint we do not
control).  the reduction is applied if we had detected no video
texturing ever happened however, despite the port having been setup
for allowing such (which is an expected situation, in fact).

there is a risk here that we may break a valid scenario if for some
reason a video texturing is engaged only following a port buffer
format change, but that is unlikely in practice (in practice, video
texturing is engaged from the start and continues accross port format
change when applicable).

however, if we had reduced the number of buffers on the port, the best
we can do is try to recover a 'clean' video texturing session on a
subsequent port format change, which is only applied if we had
established such scenario happened (ie. validated texturing scenario).

refs #SWANDROID-5661

Change-Id: Ie310a4d4d429785367b6a62e49aa5daadc401d1a
Signed-off-by: Pierre Couillaud <pierre@broadcom.com>