ir: busy wait for sys.boot.reason.
because we must launch the middleware earlier than any other
processes for various features, we also effectively need to
account for late publication of some of the settings we are
key-ing on.  in this case, the sys.boot.reason is only published
once the bootstat service comes up, so we wait.

refs #SWANDROID-5691

Change-Id: I61d8db0050228e06c35a8323bad98b9427e69df3
Signed-off-by: Pierre Couillaud <pierre.couillaud@broadcom.com>