bomx: accurate seeking improvements
Remove the timeout condition when waiting for an
updated stc sync value. This timeout was put in
place as a safety measure to prevent the decoder
from indefinitely waiting for an updated stc during
seeking. However, this situation is not likely
to occur unless the system has fallen into a
bad state. In fact, this timeout may cause the
bomx decoder to misinterpret as errors valid
use cases such as pause after seek or cases when
the player buffers too slow.

refs #SWANDROID-5714

Change-Id: Id745c9d948b37d420075e0b590b38f58e6b375f8
Signed-off-by: Luis Navarro <luis.navarro@broadcom.com>