audio: av-sync header parsing improvements
The parsing logic used to search the embedded av-headers
in the audio data  needs to be more robust, as some
content with byte patterns equal to the av-sync marker
may trigger false positives.
Two main changes to the existing logic were added:
1) When the buffer being parsed contains a fragment
of a frame which was partially written on the previous
buffer, use the pending bytes of such frame to estimate
where the next marker should be.
2) When doing blind search and a false positive occurs,
make sure to continue the search from where the false
positive occurred instead of aborting.

refs #SWANDROID-5658

Change-Id: I1ea38d9db24b5b49d93eb183606142b97126299d
Signed-off-by: Luis Navarro <luis.navarro@broadcom.com>