sec: add MediaCas proxy plug-in.
This change adds in a proxy plug-in that communicates with the
MediaCasProxyService to provide MediaCas functionality for some
CAS solutions that require networking within their plug-in.

This plug-in uses HIDL to communicate with the MediaCasProxyService
that loads the plug-ins that require networking functionality.

Please note that this is a temporary solution for CAS vendors whom
cannot initially support networking at the application layer.  This
code and the MediaCasProxy Service will be removed in a future
version.

refs #SWANDROID-5781

Change-Id: Ic9cda9d14ef125223914f0e8d115e9acc5d09008
Signed-off-by: Rob McConnell <robert.mcconnell@broadcom.com>