hals: keymaster: allow to reject some operations when "busy".
the keymaster may be invoked at any time by an application requesting
service; however in some cases where the secure hardware may be busy
with other operations, such as drm decryption, it may be preferable to
reject the keymaster operation to prevent busy usage of the secure
hardware for extended period of time which would cause deadlines on
more time sensitive secure operations to be missed.

the scheme implemented should be carefully used and not be abused, it is
however necessary in some cases.  for the time being we only reject
rsa key generation for keys of size >= 1024 bits.  more filtering of
operations may be added in the future.

refs #SWANDROID-5778

Change-Id: I502749d54dff1f79d8e632bfe721b4a4d0b29998
Signed-off-by: Pierre Couillaud <pierre.couillaud@broadcom.com>