audio: fix underruns at start of playback
On the tunneling HAL, the logic to pace the audio
data to prevent underruns was not kicking in until
certain time and buffering conditions were met.
These conditions were not taking into consideration
the timestamp of the audio data and therefore the
throttling logic was kicking in too late, resulting
sometimes in underruns at the start of playback.

refs #SWANDROID-5793

Change-Id: Ibbd421dbd526d326938df76d53f9642b36f278f3
Signed-off-by: Luis Navarro <luis.navarro@broadcom.com>