hwc2,bomx: apply geometry changes to video window faster
Make sure to apply changes to the video window geometry
whenever they occur as opposite to only when a frame is
displayed. This helps cases when there is a significant
delay between the geometry update and the next displayed
frame, for example when playing very low fps videos or
when video is paused. For cases like that it is necessary
to update the video window right away to prevent screen
artifacts.

refs #SWANDROID-5736

Change-Id: I45b9ff988b3169169d0111dfd0b0be089d6fd098
Signed-off-by: Luis Navarro <luis.navarro@broadcom.com>