prdysc: service for playready secure clock acquisition.
the playready drm service for playready 3.0 requires networking
extension to support secure clock challenge, or networking extension
is not available to standard hal's in treble architecture, so we need
to produce a service which extends the playready plugin and allows it
to issue networking requests.

that is the "prdysc" service.  the service is basic and meant for the
sole purpose of acquiring the secure clock challenge|response at the
moment.

refs #SWANDROID-5757

Change-Id: I3642bcbaa1ec5538425c62b9edf04a78fbb63b2c
Signed-off-by: Pierre Couillaud <pierre.couillaud@broadcom.com>