nxserver: rework waiter state for "secdma" (wifi) trigger condition.
there is a potential quiproquo situation in access to 'vendor/data'
during boot up of nxserver.  this in general has no side effects,
except for the case we must indeed have things right, which is the
secure dma programming for wifi.

/data/vendor is populated on post-fs-data, but nxserver is started
earlier on in post-fs; and it would effectively hold back post-fs-data
until it is done executing and triggering its waiters.  nxserver
must be started before /data in any cases since it may be required
for keymaster|gatekeeper services, such as encryption of ... /data!

so, to avoid getting in this quiproquo we must let nxserver continue
its startup while we also need to delay attempt to access /data/vendor
when absolutely necessary for the feature required. the solution is
simply to spin out a thread to wait as needed and hold back triggering
wifi driver insmod until we are ready to provide all the data it
needs (ie the secure dma information), while we let nxserver continue
its processing, allowing the system to boot no matter what.

refs #SWANDROID-5905

Change-Id: I78b6100efada8328b47a4dd35023def97de97abf
Signed-off-by: Pierre Couillaud <pierre.couillaud@broadcom.com>