bomx: prevent vdec|gralloc de-association on verified texturing.
if a true (verified) video texturing scenario is in place, one cannot
reset the vdec<->gralloc port association on port flush as this action
would void the requirement of the graphics plane with regards to keeping
the association so long the gralloc buffer mapping is alive (which it is
until a port reconfiguration takes place, a flush is not sufficient).

the side effect this is addressing are visible when playing the cts-verifier
projection video playback which has a short (30s) video which loops and
between which loop the port is flushed.  side effect is also marginally
visible on youtube 360 in some video playback, again can be traced to a
flush.

refs #SWANDROID-5806

Change-Id: I724d407f4439cef0763561c84dc3929496294c64
Fixes: SWANDROID-5485
Signed-off-by: Pierre Couillaud <pierre@broadcom.com>