hals:dpthak: re-work hfrvideo for treble
Update hfrvideo to fit better the state of the
latest Android plaform:

1) Remove the mediadrmserver process from the list
of key processes as the drm plugins don't run in
that context with the treble architecture.

2) Remove the nxserver process from the list for
hfr tunneling, as it doesn't have any impact in the
performance anymore.

3) Add the 'Niceness' elevate mode to the hfr logic.
This mode elevates the nice value of a thread instead
of elevating its scheduling policy/priority.

4) Elevate the niceness of the 'MediaCodec_loop'
threads belonging to the application. These threads
represent the start of the svp path as far as binder
transactios is concerned. By elevatings its nice value
we make sure that all subsequent binder transactios
required for svp will run at the same priority.

refs #SWANDROID-5879

Change-Id: I17cf0accd03d827937ca9bbddd8aaaa996167c28
Signed-off-by: Luis Navarro <luis.navarro@broadcom.com>