bomx: handle decoder underruns for high fps svp
Increase the minimum-buffer return watermark in
order to prevent decoder underruns as a result of
cpu starvation that may occur when handling high
fps/resolution streams. Since the cpu starvation
is more likely to occur on svp due to the larger
and more demanding video path, this change was
only enabled for secure playback at this point.

refs #SWANDROID-5879

Change-Id: I2d2dc8ad0746564553fad6af3ec1673602a5e106
Signed-off-by: Luis Navarro <luis.navarro@broadcom.com>