nxserver: disable defaultSdFormat by default
This is the equivalent of "nxserver -default_sd off", unless
composite output is enabled.

The default value is not "off", because nxserverlib initializes
the defaultSdFormat field to NEXUS_VideoFormat_eMax instead of
NEXUS_VideoFormat_eUnknown. This is to "allow cmdline or pick default
in init_session()". In our case the default was NEXUS_VideoFormat_eNtsc
as per nxserver_p_default_sd_format().

When HDMI is disconnected, this feature was automatically switching
the main display to the default SD format (NTSC), and restoring the
previous format upon HDMI reconnected.

The format change was triggering a restart of the HDCP authentication
and the QD980 HDCP test equipment did not tolerate that. The official
Simplay HDCP test equipment was fine with it though.

refs #SWSTB-8944
refs #SWANDROID-5697

Change-Id: I8a82163cf8b8a903567ef5c35c50565b575fcb1d