audio: do not use fast mixer
Netflix relis on OpenSL for UI audio.  The AudioTrack created in OpenSL
does not specify a target frameCount, resulting in the default being used.
If fast mixer is used, the resulting buffer size is too small (960 frames)
to prevent underruns when the system is busy.

This change ensures the fast mixer is not used.

refs #SWANDROID-5997

Change-Id: Id5007dbfd1b394bdc431dd053564fb3af1d316a6
Signed-off-by: Edwin Chan <edwin.chan@broadcom.com>