bomx: update stc sync constantly
It is possible that the stc sync used for
accurate seeking in tunneling, may be updated
outside a seek task by the audio hal. As it's
required by the seeking logic that bomx always
has the latest value, we need to constantly
check for changes to prevent the seek operation
from failing.

refs #SWANDROID-6102

Change-Id: Ib91cf2467bc2c0d425ceacb75726137fb04067a7
Signed-off-by: Luis Navarro <luis.navarro@broadcom.com>