bomx,audio: add audio presence indication
For video tunneling, use the shared memory
mechanism to provide a way for the audio hal
to indicate to the bomx decoder if there is
an active audio track associated with the stc.
This will be useful for the bomx video decoder
to better distinguish between audio/video or
video-only playback.

refs #SWANDROID-6106

Change-Id: Ied97ecb36ec86e73d22f810da3ef5cac89482a2c
Signed-off-by: Luis Navarro <luis.navarro@broadcom.com>
(cherry picked from commit c30d93d601e72737d7c61da0252db7c8029af030)