bomx: fine tune advertisement of video decoder resource presence.
tune the advertisement of video decoder presence to the system with the
input port state which is more accurate view of when the decoder is
being used vs just being instantiated (probed).  this in turn should
allow other parts of the system that query the decoder presence to have
more flexibility (eg keymaster) in decision.

refs #SWANDROID-6257
refs #SWANDROID-6445

Change-Id: Ic8476b96cfbef586af9d81a4b5bce911b6b33150
Signed-off-by: Pierre Couillaud <pierre.couillaud@broadcom.com>
(cherry picked from commit 30352893977c2d692aa65c2bae5141cddadf56c2)