audio: workaround contention during output open/close
There is a constant contention happening when the
framework is enumerating the hal outputs.
Opening an output sometimes fails because an output of
the same kind is already open. Most times though, the
framework is trying to close that first output while
the second one is being open, but it is unable to do
so because the open function is holding the hal lock.
When this situation happens, the open function will
just yield the hal lock for 5msec and check again
if the output in question has been closed already.

refs #SWANDROID-5915

Change-Id: Ibf00f7458ab7c7e729da889c7985326cc89628de
Signed-off-by: Luis Navarro <luis.navarro@broadcom.com>
(cherry picked from commit adbc0bf4007e5d6325ff003f860f89a6642bf210)