bomx: move picture heap toggling-out outside constructor
When there is widevine decryption in progress, sage will
invalidate current key slots when the picture heap is toggled
from secure to open. Even a simple codec enumeration done by
the framework may trigger this scenario, which would result
in existing widevine decryption keys to be invalidated and
consequently cause decryption failures.

This change attempts to alleviate that by moving the toggling
secure->open outside the constructor and have it instead in
the loaded->idle transition, which can be interpreted as an
early indication of an intent to do decoding.

refs #SWANDROID-6420

Change-Id: I7a16b24e63bc0223cc57fffc72f53f56b2ac9a9a
Signed-off-by: Luis Navarro <luis.navarro@broadcom.com>
(cherry picked from commit 58c2be5725c7f0d54817ce33268739e4d046bcd8)